airflow_local/requirements.txt
# airflow_local/requirements.txt
apache-airflow-providers-google
pandas
pyarrow
nba_api
requests